requirements.txt
pandas
numpy
streamlit 
sentence-transformers 
faiss-cpu
matplotlib
seaborn
ipykernel
requests
fastapi
uvicorn
pandas
sentence-transformers
faiss-cpu
numpy
python-multipart